Moving splitter to separate class

+++ src/main/java/com/tgelder/downhill/rngs/RNG.java
@@ -2,6 +2,6 @@ package com.tgelder.downhill.rngs;
 
 public interface RNG {
   
-  public float getNext();
+  public double getNext();
 
 }
