+++ src/main/java/com/tgelder/downhill/rngs/RandomRNG.java
@@ -11,7 +11,7 @@ public class RandomRNG implements RNG {
   }
 
   @Override
-  public float getNext() {
+  public double getNext() {
     return random.nextFloat();
   }
